Moved the comm port interactions to CameraHandler
